Move configuration related functions into its own module
